Enforce backward-compatibility when encoding NominalDiffTime.
  Keep encoding as integer when the value is a whole number of seconds. Otherwise, this may break clients down the line.

+++ server/test/unit/Ogmios/Data/JsonSpec.hs
@@ -385,7 +385,7 @@ spec = do
                     \7A80D87A801A002625A0D87983D879801A000F4240D879811A000FA92E"
 
     context "SlotLength" $ do
-        let matrix = [ ( mkSlotLength 1, Json.double 1 )
+        let matrix = [ ( mkSlotLength 1, Json.integer 1 )
                      , ( mkSlotLength 0.1, Json.double 0.1 )
                      ]
         forM_ matrix $ \(slotLength, json) ->
